+++ includes/admin/views/view-tools-cache.php
@@ -36,6 +36,7 @@
                 'name' => __('Name', 'oes'),
                 'id' => __('Key', 'oes'),
                 'type' => __('Type', 'oes'),
+                'cache_language' => __('Language', 'oes'),
                 'parts' => __('Parts', 'oes'),
                 'size' => __('Size (KB)', 'oes'),
                 'timestamp' => __('Timestamp', 'oes')